app/requirements.txt
uvicorn==0.12.2
pydantic==1.6.1
email-validator==1.1.2
debugpy==1.0.0
requests==2.24.0
psycopg2==2.8.6
redis==3.5.3
peewee==3.13.3
python-multipart==0.0.5
fastapi==0.61.1
pytest==6.1.1
bcrypt==3.2.0
cryptography==3.2.1
python-jose==3.2.0
passlib==1.7.4